split up tests to remove complexity from test-skipping
